requirements.txt
openai>=2.0,<3.0
llama-index-core>=0.12,<0.13
llama-index-llms-openai>=0.3,<0.4
pandas>=2.2,<3.0
streamlit>=1.45,<2.0
